Format corrections
Reviewed all text files for each book for formatting errors and corrected as needed.

Changed the symbols for the 2nd Alternative word in Hebrew OT books too now use a flower punctuation mark instead of a double * to aviod letter overlap.

Changed layout when viewed on a small screen or mobile device to give an easier utility.