Search options and Google Translate
Added the ability to automatically remove one of the word in discrepency when doing a search, and also the ability to clear any previously applied mark up to the text formatting.
Also added a Google Translate feature to the header to help translate the page contents.
Updated help file instructions.